requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn==1.2.2
pycaret[full]
azure-storage-blob
openai
python-dotenv